+++ src/helper/util.js
@@ -37,7 +37,7 @@ const generateEmployeeChoices = (employeesFromDB) => {
 module.exports = {
   verifyNumber,
   verifyResponses,
-  generateDepartmentChoices,
   generateRoleChoices,
+  generateDepartmentChoices,
   generateEmployeeChoices,
 };
